+++ src/wildboar/distance/__init__.py
@@ -24,6 +24,7 @@ from ._distance import (
     subsequence_match,
 )
 from ._matrix_profile import matrix_profile
+from ._neighbour import KNeighbourClassifier
 
 __all__ = [
     "pairwise_subsequence_distance",
@@ -33,5 +34,6 @@ __all__ = [
     "pairwise_distance",
     "paired_distance",
     "matrix_profile",
+    "KNeighbourClassifier",
     "mean_paired_distance",  # Deprecated
 ]